[#99] Adding structural implementation
